fix: Linux driver and memory mapping issues
I swapped the entire mem_module by @treecodeXXI out for my own
implementation, because it was simpler and more tidy, plus I was already
familiar with it, so I had more hope in getting it to work.

My implementation, however, uses a device file approach, not a sysfs
approach, which means that now the memory region is accessible through
/dev/simics_shm, rather than /sys/kernel/mem_dev/mem.

On that note, I've renamed mem_dev and the entire module to simics_shm,
because it's a more fitting name.

The other half of this commit is fixing map_memory.simics. It was a long
search, but instantiating our own class was a mistake, because we want
to have RAM instead.

I believe my changes to risc-v-simple.dts are irrelevant and did not
change the behavior of anything, but feel like a worthy addition.
It is worth noting that, in the context of the dts, "simics_shm" refers
to the generic memory region created and governed by Simics, NOT the
kernel module of the same name. The "simics_shm" kernel module merely
relies upon this memory existing on the given physical address.

Also included a minor shell.nix fix for buildroot menuconfig.